Adding email & posts fields to the User model & adding the Post model, setting its relations with User and CRUD mutations

+++ graphql-prisma/src/schema/user.ts
@@ -5,14 +5,22 @@ export const User = objectType({
   definition(t) {
     t.model.id()
     t.model.name()
+    t.model.email()
+    t.model.posts()
   },
 })
 
 export const createUser = mutationField('createUser', {
   type: User,
-  args: { name: nonNull(stringArg()) },
-  resolve(_root, { name }, { prisma }) {
-    return prisma.user.create({ data: { name } })
+  args: { name: nonNull(stringArg()), email: nonNull(stringArg()) },
+  async resolve(_root, { name, email }, { prisma }) {
+    const emailTaken = await prisma.user.findUnique({ where: { email } })
+
+    if (emailTaken) {
+      throw new Error('An account is already using this email')
+    }
+
+    return prisma.user.create({ data: { name, email } })
   },
 })
 
@@ -21,7 +29,15 @@ export const deleteUser = mutationField('deleteUser', {
   args: {
     id: nonNull(idArg()),
   },
-  resolve(_root, { id }, { prisma }) {
+  async resolve(_root, { id }, { prisma }) {
+    const userExists = await prisma.user.findUnique({ where: { id } })
+
+    if (!userExists) {
+      throw new Error('User not found')
+    }
+
+    await prisma.post.deleteMany({ where: { userId: id } })
+
     return prisma.user.delete({ where: { id } })
   },
 })
@@ -30,16 +46,30 @@ export const updateUser = mutationField('updateUser', {
   type: User,
   args: {
     whereId: nonNull(idArg()),
-    updateName: 'String',
+    updateName: stringArg(),
+    updateEmail: stringArg(),
   },
-  resolve(_root, { whereId, updateName }, { prisma }) {
+  async resolve(_root, { whereId, updateName, updateEmail }, { prisma }) {
+    const userExists = await prisma.user.findUnique({ where: { id: whereId } })
+
+    if (!userExists) {
+      throw new Error('User not found')
+    }
+
+    let data: { email?: string; name?: string } = {}
+
+    if (updateEmail) {
+      data.email = updateEmail
+    }
+
     if (updateName) {
-      return prisma.user.update({
-        where: { id: whereId },
-        data: { name: updateName },
-      })
-    } else {
+      data.name = updateName
+    }
+
+    if (!updateName && !updateEmail) {
       throw new Error('Please provide something to update')
     }
+
+    return prisma.user.update({ where: { id: whereId }, data })
   },
 })
